Done some work to ensure the message sent matches the case sensitivity, Regardless if shift was down or capslock was down.

+++ JavaMacro/src/macro/win32/inferfaces/KbInterface.java
@@ -25,7 +25,7 @@ public interface KbInterface extends User32, StdCallLibrary {
 
     boolean SetKeyboardState(byte[] lpKeyState);
 
-    int VkKeyScan(char ch);
+    short VkKeyScan(char ch);
 
     short GetAsyncKeyState( int i);
